feat: :sparkles: prod action for publishing to main
