+++ github-action/src/releasePr.ts
@@ -1,15 +1,16 @@
 #!/usr/bin/env node
 
 import { Command } from 'commander';
-import fs from 'fs';
-import { execSync } from 'child_process';
-import { parsePackageJson } from './parsePackageJson.js';
-import { resolve } from 'path';
-import AdmZip from 'adm-zip';
 import { HttpClient } from '@actions/http-client';
 import { OutgoingHttpHeaders } from 'http';
-import { Readable } from 'stream';
-import { notice, setSecret } from '@actions/core';
+import { setSecret } from '@actions/core';
+import { parsePackageJson } from './utils/parsePackageJson.js';
+import { prepareBundle } from './utils/prepareBundle.js';
+import { makeManifest } from './utils/makeManifest.js';
+import { zipBundle } from './utils/zipBundle.js';
+import { uploadBundle } from './utils/uploadBundle.js';
+
+const ciUrl = 'https://fusion-s-portal-ci.azurewebsites.net';
 
 const program = new Command();
 
@@ -30,16 +31,10 @@ program
 await program.parseAsync();
 
 export async function release(token: string, prNumber: string) {
-  // Vite build
-  notice('bundling application');
   prepareBundle();
 
-  // Create manifest
-  notice('making manifest');
   makeManifest('./package.json');
 
-  // zip bundle
-  notice('zipping bundle');
   const zipped = zipBundle();
 
   const r = parsePackageJson();
@@ -49,106 +44,8 @@ export async function release(token: string, prNumber: string) {
     );
   }
 
-  await uploadBundle(token, r.name, zipped, prNumber);
-}
-
-async function uploadBundle(
-  token: string,
-  appKey: string,
-  zipped: AdmZip,
-  prNumber: string
-) {
-  const client = new HttpClient();
-
-  const headers: OutgoingHttpHeaders = {
-    ['Authorization']: `Bearer ${token}`,
-    ['Content-Type']: 'application/zip',
-    ['Content-Disposition']: 'attachment; filename=bundle.zip',
-  };
-
-  const stream = Readable.from(zipped.toBuffer());
-
-  const r = await client.sendStream(
-    'POST',
-    `https://fusion-s-portal-ci.azurewebsites.net/api/apps/${appKey}/versions`,
-    stream,
-    headers
-  );
-
-  notice(`bundle uploaded with status code ${r.message.statusCode}`);
-  if (r.message.statusCode !== 200) {
-    throw new Error('Bundle failed to upload, fatal error');
-  }
-
-  /** Publish bundle */
-  const publishResponse = await client.post(
-    `https://fusion-s-portal-ci.azurewebsites.net/api/apps/${appKey}/publish`,
-    '',
-    headers
-  );
-  if (publishResponse.message.statusCode !== 200) {
-    throw new Error(JSON.stringify(publishResponse.message));
-  }
-
-  await patchWithPrNumber(prNumber, token, appKey);
-}
-
-function prepareBundle() {
-  const { name } = parsePackageJson();
-  if (!name) {
-    throw new Error('Missing name in package.json');
-  }
-  execSync('npx vite build --logLevel silent', { stdio: 'inherit' });
-}
-
-export function makeManifest(path: string) {
-  const { version, name, ...maybe } = parsePackageJson(path);
-  if (!version || !name) {
-    throw new Error('Name or version missing in package.json');
-  }
-  const { major, minor, patch } = splitVersions(version);
-
-  /** Some app-manifests have custom short and displaynames */
-  const shortName = maybe?.['shortName'] ?? name;
-  const displayName = maybe?.['displayName'] ?? name[0].toUpperCase() + name.slice(1);
-
-  const manifest = {
-    name: displayName,
-    shortName: shortName,
-    key: name,
-    version: {
-      major: major,
-      minor: minor,
-      patch: patch,
-    },
-  };
-
-  const data = JSON.stringify(manifest, null, 2);
-
-  fs.writeFileSync('./dist/app-manifest.json', data);
-}
-
-function splitVersions(version: string) {
-  const [major, minor, patch] = version.split('.');
-  return {
-    major,
-    minor,
-    patch,
-  };
-}
-
-export function zipBundle() {
-  const appManifestPath = resolve('./dist/app-manifest.json');
-  const bundlePath = resolve('./dist/app-bundle.js');
-
-  var zip = new AdmZip();
-
-  //TODO: scan files in package.json
-  zip.addLocalFile(appManifestPath);
-  zip.addLocalFile(bundlePath);
-
-  zip.writeZip('./dist/bundle.zip');
-  return zip;
+  await uploadBundle(ciUrl, token, r.name, zipped);
+  await patchWithPrNumber(prNumber, token, r.name);
 }
 
 async function patchWithPrNumber(prNumber: string, token: string, appKey: string) {
@@ -160,10 +57,7 @@ async function patchWithPrNumber(prNumber: string, token: string, appKey: string
   };
 
   //Download current config
-  const res = await client.get(
-    `https://fusion-s-portal-ci.azurewebsites.net/api/apps/${appKey}/config`,
-    headers
-  );
+  const res = await client.get(`${ciUrl}/api/apps/${appKey}/config`, headers);
   if (res.message.statusCode !== 200) {
     throw new Error('Failed to fetch client config');
   }
@@ -173,7 +67,7 @@ async function patchWithPrNumber(prNumber: string, token: string, appKey: string
 
   //patch
   const patchResponse = await client.put(
-    `https://fusion-s-portal-ci.azurewebsites.net/api/apps/${appKey}/config`,
+    `${ciUrl}/api/apps/${appKey}/config`,
     JSON.stringify(config),
     headers
   );
