+++ github-action/src/utils/parsePackageJson.ts
