+++ github-action/src/utils/prepareBundle.ts
@@ -1,0 +1,7 @@
+import { execSync } from 'child_process';
+import { notice } from '@actions/core';
+
+export function prepareBundle() {
+  notice('bundling application');
+  execSync('npx vite build --logLevel silent', { stdio: 'inherit' });
+}
